Module Text_block

type t

Two dimensional blocks of text

val sexp_of_t : t -> Sexplib0.Sexp.t

include Core.Invariant.S with type t := t

val invariant : t -> unit

val nil : t

The empty block. a left and right unit to both hcat and vcat

val fill : char -> width:int -> height:int -> t

fill and space assume width and height are non-negative

val space : width:int -> height:int -> t

val fill_uchar : Core.Uchar.t -> width:int -> height:int -> t

Fill a space with a Unicode character

type valign = [

| `Top

| `Bottom

| `Center

]

Vertical and horizontal alignment specifications

type halign = [

| `Left

| `Right

| `Center

]

val text : ?align:halign -> ?max_width:int -> string -> t

A basic block of text, split on newlines and horizontally aligned as specified.

If max_width is provided, split each line of the input on whitespace and wrap words to respect the request. So long as no words are longer than max_width, the resulting text block will be no wider than max_width

val textf : ?align:halign -> ?max_width:int -> ( 'r, unit, string, t ) Core.format4 -> 'r

Like text, but takes a format string like printf

val vcat : ?align:halign -> ?sep:t -> t list -> t

Vertical concatenation with alignment

val hcat : ?align:valign -> ?sep:t -> t list -> t

Horizontal concatenation with alignment

Text block dimensions

val width : t -> int

val height : t -> int

Vertical and horizontal sequence alignment. Both valign and halign return a list of the same length as the input, with the corresponding elements padded to the appropriate alignment.

If you have a list of a statically known length, using With_static_lengths.valign or With_static_lengths.halign below will let the type checker know that the length of the returned list is equal to the length of the input list.

val valign : valign -> t list -> t list

val halign : halign -> t list -> t list

Empty blocks with either horizontal or vertical extent -- useful for specifying a minimum width or height in conjunction with valign or halign, respectively

val hstrut : int -> t

val vstrut : int -> t

val ansi_escape : ?prefix:string -> ?suffix:string -> t -> t

Wrap a block with an ANSI escape sequence. The prefix and suffix arguments should render with zero width and height.

val render : t -> string

render a block of text as a string

val table : ?sep_width:int -> [ `Cols of (t list * halign) list ] -> [ `Rows of t list ]

Alignment of a 2D grid of blocks

val compress_table_header : ?sep_width:int -> [ `Cols of (t * t list * halign) list ] -> [ `Header of t ] * [ `Rows of t list ]

Compress table header according to column widths.

Input: a list of columns of the form (title, values, column alignment). Output: one header block and row sequence. Raises: if the values lists are not the same length in each column. Example:

                                                 first name
      age  first name  last name            age  |     last name
      |    |           |            ==>     |    |     |
      30   sue         smith                30   sue   smith
      18   bill        rodriguez            18   bill  rodriguez
      76   rick        jones                76   rick  jones

module Boxed : sig ... end

Combinators for building up cell structures separated by box characters: e.g.

val boxed : Boxed.t -> t

See comment for Boxed

val vsep : t

vsep = vstrut 1

val hsep : t

hsep = hstrut 1

val indent : ?n:int -> t -> t

indent ~n t = hcat [hstrut n; t]. n defaults to 2

val sexp : ( 'a -> Core.Sexp.t ) -> 'a -> t

sexp sexp_of_a a = sexp_of_a a |> Sexp.to_string |> text

module With_static_lengths : sig ... end

Versions of halign and valign with the invariant about list length encoded into the types.